For 2D translation motion-correction, start with the middle recording cycle
Assuming the recordings are supplied in chronological order, it makes more sense
to motion correct starting from the middle and working our way out to the start
and end of a recording session. This is because there will be some drift in the
field of view over the course of a recording session, and the middle of the
recording is likely to be closer an optimal alignment template. We expect there
to be less displacement from the middle to either start or end than there is
between the start and end trials.

I have left the loop through a single cycle going in chronological order for
serial and random order for parallel processing. Unless there is a small number
of very long cycles, this order is expected to have less impact on the final
result.