Fix problem with ROIs drawn outside of image boundary
The ROI coordinates are stored as signed (not unsigned) integers,
but the `imagej.read_roi` function loading method was only valid
for unsigned integers.

This was causing a problem specifically when ROIs were drawn
outside of the image (which is possible in ImageJ by dragging a
box outside the window edge). In this case, the top or left coordinate
values would be returned as very large positive values.

This problem is fixed, but I don't know which other values in the
ImageJ file are stored as signed as opposed to unsigned integers, so
other values may still be loaded incorrectly.